Guess_the_number
This file consists of code for "GUESS THE NUMBER" game using Java. Through this code user can generate the random number and compare the user input with random number, loops are used to give  the user input until the user input = random number. At last, it also gives you the number of attempts to guess the correct number.